requirements.txt
tzlocal~=4.2

Django~=4.1.7
djangorestframework~=3.14.0

psycopg2-binary
Pillow

drf-yasg
django-filter
dj-rest-auth

python-decouple